Finish BigDecimals transfer
